Convert remaining functions
