+++ frontend/src/metabase/home/homepage/utils.ts
@@ -1,4 +1,4 @@
-import moment, { MomentInput } from "moment";
+import moment, { MomentInput } from "moment-timezone";
 import { parseTimestamp } from "metabase/lib/time";
 
 export const isWithinWeeks = (
